Added default LIRCD configuration
Only speculative, given that I've left the robot in school and so can't actually test it, but... I'm seeing issues using the remote when I run the robot from cold in a "live" environment. I'm thinking that's because the lircd process isn't being given a default remote configuration (which correlates to the fact that all seems fine when I test using the lirc-robot.py program) - so now I am.